+++ openbb_terminal/core/session/session_model.py
@@ -21,7 +21,10 @@ from openbb_terminal.core.session.current_user import (
     set_current_user,
     set_default_user,
 )
-from openbb_terminal.core.session.routines_handler import download_and_save_routines
+from openbb_terminal.core.session.routines_handler import (
+    download_routines,
+    save_routine,
+)
 from openbb_terminal.core.session.sources_handler import get_updated_hub_sources
 from openbb_terminal.core.session.utils import run_thread
 from openbb_terminal.helper_funcs import system_clear
@@ -103,20 +106,10 @@ def login(session: dict) -> LoginStatus:
             set_current_user(hub_user)
 
             auth_header = hub_user.profile.get_auth_header()
-            updated_sources = get_updated_hub_sources(configs)
-            if updated_sources:
-                run_thread(
-                    Hub.upload_user_field,
-                    {
-                        "key": "features_sources",
-                        "value": updated_sources,
-                        "auth_header": auth_header,
-                        "silent": True,
-                    },
-                )
             run_thread(
-                download_and_save_routines, {"auth_header": auth_header, "silent": True}
+                update_backend_sources, {"auth_header": auth_header, "configs": configs}
             )
+            run_thread(download_and_save_routines, {"auth_header": auth_header})
 
             Local.apply_configs(configs)
             Local.update_flair(get_current_user().profile.username)
@@ -126,6 +119,50 @@ def login(session: dict) -> LoginStatus:
             return LoginStatus.UNAUTHORIZED
         return LoginStatus.FAILED
     return LoginStatus.NO_RESPONSE
+
+
+def update_backend_sources(auth_header, configs, silent: bool = True):
+    """Update backend sources if new source or command path available.
+
+    Parameters
+    ----------
+    auth_header : str
+        The authorization header, e.g. "Bearer <token>".
+    configs : Dict
+        Dictionary with configs
+    silent : bool
+        Whether to silence the console output, by default True
+    """
+    console_print = console.print if not silent else lambda *args, **kwargs: None
+
+    try:
+        updated_sources = get_updated_hub_sources(configs)
+        if updated_sources:
+            Hub.upload_user_field(
+                key="features_sources",
+                value=updated_sources,
+                auth_header=auth_header,
+                silent=silent,
+            )
+    except Exception:
+        console_print("[red]Failed to update backend sources.[/red]")
+
+
+def download_and_save_routines(auth_header: str, silent: bool = True):
+    """Download and save routines.
+
+    Parameters
+    ----------
+    auth_header : str
+        The authorization header, e.g. "Bearer <token>".
+    silent : bool
+        Whether to silence the console output, by default True
+    """
+    routines = download_routines(auth_header=auth_header, silent=silent)
+    for name, content in routines.items():
+        save_routine(
+            file_name=f"{name}.openbb", routine=content, force=True, silent=silent
+        )
 
 
 def logout(
